aircraft_inspection/requirements.txt
absl-py==1.4.0
accelerate==0.21.0
addict==2.4.0
aenum==3.1.16
aio-pika==8.2.3
aiofiles==23.2.1
aiogram==2.15
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiohttp-retry==2.8.3
aiormq==6.4.2
aiosignal==1.3.1
albumentations==1.4.3
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
APScheduler==3.9.1.post1
asgiref==3.9.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.15.0
bidict==0.23.1
blendmodes==2022
blinker==1.9.0
boto3==1.34.133
botocore==1.34.133
CacheControl==0.12.14
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
colorlog==6.9.0
comm==0.2.2
comtypes==1.4.10
confluent-kafka==2.4.0
controlnet_aux==0.0.10
cryptography==42.0.8
cssselect2==0.8.0
cycler==0.12.1
Cython==3.1.1
dask==2022.10.2
DateTime==5.5
debugpy==1.8.13
decorator==5.2.1
deprecation==2.1.0
depth_anything @ https://github.com/huchenlei/Depth-Anything/releases/download/v1.0.0/depth_anything-2024.1.22.0-py2.py3-none-any.whl#sha256=26c1d38b8c3c306b4a2197d725a4b989ff65f7ebcf4fb5a96a1b6db7fbd56780
depth_anything_v2 @ https://github.com/MackinationsAi/UDAV2-ControlNet/releases/download/v1.0.0/depth_anything_v2-2024.7.1.0-py2.py3-none-any.whl#sha256=6848128867d1f7c7519d88df0f88bfab89100dc5225259c4d7cb90325c308c9f
dialogflow==2.0.0
diskcache==5.6.3
Django==5.2.4
dnspython==2.3.0
docopt==0.6.2
dsine @ https://github.com/sdbds/DSINE/releases/download/1.0.2/dsine-2024.3.23-py3-none-any.whl#sha256=b9ea3bacce09f9b3f7fb4fa12471da7e465b2f9a60412711105a9238db280442
easydict==1.13
easygui==0.98.3
einops==0.4.1
exceptiongroup==1.2.2
executing==2.2.0
facexlib==0.3.0
fastapi==0.94.0
fbmessenger==6.0.0
ffmpy==0.5.0
filelock==3.18.0
filterpy==1.4.5
fire==0.6.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Mail==0.10.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.4.0
geffnet==1.0.2
gitdb==4.0.12
GitPython==3.1.32
glob2==0.5
google-api-core==1.34.1
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
gotrue==2.12.0
gradio==3.41.2
gradio_client==0.5.0
greenlet==3.0.3
groovy==0.1.2
grpcio==1.64.1
grpcio-status==1.48.2
h11==0.12.0
h2==3.2.0
h5py==3.11.0
handrefinerportable @ https://github.com/huchenlei/HandRefinerPortable/releases/download/v1.0.1/handrefinerportable-2024.2.12.0-py2.py3-none-any.whl#sha256=1e6c702905919f4c49bcb2db7b20d334e8458a7555cd57630600584ec38ca6a9
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.15.0
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
ifnude==0.0.3
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.1.0
insightface @ https://github.com/Gourieff/Assets/raw/main/Insightface/insightface-0.7.3-cp310-cp310-win_amd64.whl#sha256=47aa0571b2aadd8545d4bc7615dfbc374c10180c283b7ac65058fcb41ed4df86
iopath==0.1.9
ipykernel==6.29.5
ipython==8.34.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.2.0
jsonmerge==1.8.0
jsonpickle==3.0.4
jsonschema==4.17.3
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.12.0
kiwisolver==1.4.5
kornia==0.6.7
lark==1.1.2
lazy_loader==0.4
libclang==18.1.1
lightning-utilities==0.14.3
llvmlite==0.44.0
locket==1.0.0
lxml==5.4.0
manifold3d==3.1.1
mapbox_earcut==1.0.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mattermostwrapper==2.2
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==5.2.0
narwhals==1.41.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.2.3
onnx==1.18.0
onnxruntime==1.22.0
onnxruntime-gpu==1.22.0
open-clip-torch==2.20.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
orjson==3.10.16
packaging==23.1
pamqp==3.2.1
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pefile==2023.2.7
piexif==1.1.3
Pillow==9.5.0
pillow-avif-plugin==1.4.3
platformdirs==4.3.7
pluggy==1.5.0
portalocker==2.10.0
postgrest==1.0.2
prettytable==3.16.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==3.20.0
psutil==5.9.5
psycopg2-binary==2.9.9
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycollada==0.9
pycparser==2.22
pydantic==1.10.17
pydantic_core==2.33.1
pydot==1.4.2
pydub==0.25.1
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.10.1
pykwalify==1.8.0
pymongo==4.3.3
pyparsing==3.1.2
pyperclip==1.9.0
pypiwin32==223
pyreadline3==3.4.1
pyrsistent==0.20.0
pytest==8.3.5
pytest-mock==3.14.1
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-engineio==3.13.2
python-multipart==0.0.20
python-socketio==4.6.1
pytorch-lightning==1.9.4
pyttsx3==2.98
pytz==2022.7.1
PyWavelets==1.8.0
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.1
pyzmq==26.3.0
questionary==1.10.0
randomname==0.1.5
rasa==3.6.20
rasa-sdk==3.6.2
realtime==2.4.3
redis==4.6.0
regex==2022.10.31
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resize-right==0.0.2
rfc3986==1.5.0
rich==14.0.0
rocketchat-API==1.30.0
rsa==4.9
rtree==1.4.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.8
ruff==0.11.6
s3transfer==0.10.2
safehttpx==0.1.6
safetensors==0.4.2
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-image==0.21.0
scikit-learn==1.7.0
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==1.14.0
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.30.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
spandrel==0.3.4
spandrel_extra_arches==0.1.1
SpeechRecognition==3.14.2
SQLAlchemy==1.4.52
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.26.1
storage3==0.11.3
StrEnum==0.4.15
structlog==23.3.0
structlog-sentry==2.1.0
supafunc==0.9.4
svg.path==6.3
svglib==1.5.1
sympy==1.14.0
tabulate==0.9.0
tarsafe==0.0.4
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.13.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2025.5.10
timm==0.9.5
tinycss2==1.4.0
tokenizers==0.13.3
tomesd==0.1.3
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
toolz==0.12.1
torch==2.1.2+cu121
torchdiffeq==0.2.3
torchmetrics==1.7.2
torchsde==0.2.6
torchvision==0.16.2+cu121
tornado==6.4.2
tqdm==4.66.4
traitlets==5.14.3
trampoline==0.1.2
transformers==4.30.2
trimesh==4.6.11
twilio==8.2.2
typer==0.15.2
typing-inspection==0.4.0
typing-utils==0.1.0
typing_extensions==4.13.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
ultralytics==8.3.168
ultralytics-thop==2.0.14
urllib3==2.2.2
uvicorn==0.34.1
vhacdx==0.0.8.post2
wcwidth==0.2.13
webencodings==0.5.1
webexteamssdk==1.6.1
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.14.1
wsproto==1.2.0
xformers==0.0.23.post1
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.20.0
zipp==3.22.0
zope.interface==7.2
